checkrunner-back/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
boto3==1.17.84
botocore==1.20.84
click==8.0.1
colorama==0.4.4
coverage==5.5
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.12.1
mistune==0.8.4
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-cov==2.12.0
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.4.2
six==1.16.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
